Move more tests to CalciteJoinQueryTest
